% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstree-selector-input.R
\name{jsTreeSelectorInput}
\alias{jsTreeSelectorInput}
\title{R Shiny Input using a jsTree to select a node}
\usage{
jsTreeSelectorInput(inputID, df = NULL, selected = NULL)
}
\arguments{
\item{inputID}{The unique id to set for this element}

\item{df}{A dataframe with 'id', 'text', 'parent', and 'value' fields}

\item{select}{The id of the node to default as selected (unless id is not used, then it is the text)}
}
\description{
A reactive input value based on clicks on the jsTree nodes.
}
