% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-brush-plot.R
\name{polygonBrushPlot}
\alias{polygonBrushPlot}
\title{R Shiny Plot With Additional Polygon Brush}
\usage{
polygonBrushPlot(inputID)
}
\arguments{
\item{inputID}{The unique id to set for this element}
}
\description{
A reactive input value based on clicks on the jsTree nodes.
}
