% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstree-selector-input.R
\name{updatejsTreeSelectorInput}
\alias{updatejsTreeSelectorInput}
\title{update an existing jsTreeSelectorInput}
\usage{
updatejsTreeSelectorInput(session, inputID, df, selected = NULL)
}
\arguments{
\item{session}{The session this input is part of}

\item{inputID}{The unique id to set for this element}

\item{df}{A dataframe to replace the current one with 'id', 'text', 'parent', and 'value' fields}

\item{select}{The id of the node to default as selected (unless id is not used, then it is the text)}
}
\description{
A reactive input value based on clicks on the jsTree nodes.
}
